Create Movie recommendation using ALS
Movielens100k dataset, which is a set of 100,000 data points related to ratings given by a set of users to a set of movies. The dataset consists of user ID, movie ID, rating, and timestamp fields. Alternating least squares (ALS), which solves matrix factorization problems, is used to train the model.